+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # XRAYUI Changelog
 
+## [0.32.0] - 2025-02-02
+
+> _Important: Please clear your browser cache (e.g.**ctrl+f5**) to ensure that outdated files are updated._
+
+- ADDED: When importing a new client configuration, it is now possible to select the services to unblock. They will automatically be added to the `routing rules`, so there's no need to set them up manually later.
+- ADDED: A checkbox `Don't break my network devices` in the `import configuration` window. This will set the default mode to `bypass`, potentially preventing disruption to `IoT` devices on the network.
+- FIXED: Correctly display advanced `DNS` server name when defined.
+- FIXED: The `bypass`/`redirect` mode dropdown value sometimes remains unchanged.
+- FIXED: a bug with loading display function was not resolved.
+- REMOVED: Unused/abandoned parts of the code.
+- IMPROVED: Mounting of the menu during installation.
+- IMPROVED: Minor cosmetic fixes and changes.
+
 ## [0.31.1] - 2025-01-30
 
 > _Important: Please clear your browser cache (e.g.**ctrl+f5**) to ensure that outdated files are updated._